+++ test/python/circuit/test_controlled_gate.py
@@ -16,7 +16,7 @@
 import unittest
 from test import combine
 import numpy as np
-from numpy import pi
+from numpy import pi, sin, cos
 from ddt import ddt, data, unpack
 
 from qiskit import QuantumRegister, QuantumCircuit, execute, BasicAer, QiskitError
@@ -76,6 +76,7 @@ from qiskit.circuit.library import (
 from qiskit.circuit._utils import _compute_control_matrix
 import qiskit.circuit.library.standard_gates as allGates
 from qiskit.extensions import UnitaryGate
+from qiskit.circuit.library.standard_gates.multi_control_rotation_gates import linear_depth_mcv
 
 from .gate_utils import _get_free_params
 
@@ -576,6 +577,18 @@ class TestControlledGate(QiskitTestCase):
         base = XGate().to_matrix()
         expected = _compute_control_matrix(base, num_controls)
         self.assertTrue(matrix_equal(simulated, expected, atol=1e-8))
+
+    def test_linear_depth_mcv(self):
+        """Test linear_depth_mcv"""
+        num_ctrls = 6
+        theta = 0.3
+        qc = QuantumCircuit(num_ctrls + 1)
+        x = -sin(theta / 2)
+        z = cos(theta / 2)
+        linear_depth_mcv(qc, x, z, list(range(num_ctrls)), num_ctrls)
+        ry_matrix = RYGate(theta).to_matrix()
+        mcry_matrix = _compute_control_matrix(ry_matrix, 6)
+        self.assertTrue(np.allclose(mcry_matrix, Operator(qc).to_matrix()))
 
     @combine(num_controls=[1, 2, 4], base_gate_name=["x", "y", "z"], use_basis_gates=[True, False])
     def test_multi_controlled_rotation_gate_matrices(
